plc_header_generator::header_generator::file_helper

Trait FileHelper

pub trait FileHelper {
    // Required methods
    fn get_directory(&self) -> &str;
    fn set_directory(&mut self, directory: &str);
    fn get_path(&self) -> &str;
    fn set_path(&mut self, path: &str);
    fn get_file_name(&self) -> &str;
    fn set_file_name(&mut self, file_name: &str);
    fn get_formatted_path(&self) -> &str;
    fn set_formatted_path(&mut self, formatted_path: &str);
    fn determine_header_file_information(
        &mut self,
        generate_header_options: &GenerateHeaderOptions,
        compilation_unit: &CompilationUnit,
    ) -> bool;
}

Required Methods

fn get_directory(&self) -> &str

Returns the directory the header should be written to

This should return the directory for writing the header to (without the header file). It can return an empty string if the header is being written to the same directory as the interface itself.

fn set_directory(&mut self, directory: &str)

Sets the directory the header should be written to

fn get_path(&self) -> &str

Returns the file path that the header file should be written to

fn set_path(&mut self, path: &str)

Sets the file path that the header file should be written to

fn get_file_name(&self) -> &str

fn set_file_name(&mut self, file_name: &str)

Sets the file name for the header file

fn get_formatted_path(&self) -> &str

fn set_formatted_path(&mut self, formatted_path: &str)

Sets the formatted path for the header file

fn determine_header_file_information( &mut self, generate_header_options: &GenerateHeaderOptions, compilation_unit: &CompilationUnit, ) -> bool

Determines file information for the header file and returns whether or not the determination was successful

The succesful result of this method must be that the “directory” (accessible via the “get_directory” method) and the “path” (accessible via the “get_path” method) are both populated with valid results.