plc_header_generator::header_generator

Function prepare_template_data_for_header_generation

pub fn prepare_template_data_for_header_generation(
    generate_header_options: &GenerateHeaderOptions,
    compilation_unit: &CompilationUnit,
) -> Result<Box<dyn GeneratedHeader>, Diagnostic>

Prepares the header data for generation given the options for generation and a compilation unit.

Should the process fail to determine an output directory and path for the header file a none-generated header will be returned, this is a valid outcome as some compilation units are not related to a file.