plc_header_generator::header_generator::template_helper

Trait TemplateHelper

pub trait TemplateHelper {
    // Required methods
    fn get_template_data(&self) -> &TemplateData;
    fn set_template_data(&mut self, template_data: TemplateData);
    fn get_mutable_template_data_user_defined_types(
        &mut self,
    ) -> &mut UserDefinedTypes;
    fn get_mutable_template_data_global_variables(
        &mut self,
    ) -> &mut Vec<Variable>;
    fn get_mutable_template_data_functions(&mut self) -> &mut Vec<Function>;
    fn get_template(&self, template_type: TemplateType) -> Template;
}

Required Methods

fn get_template_data(&self) -> &TemplateData

Returns the template data of the header file, which is an object representation of the data that will be written to the header template

fn set_template_data(&mut self, template_data: TemplateData)

Sets the template data on the header

fn get_mutable_template_data_user_defined_types( &mut self, ) -> &mut UserDefinedTypes

Returns the user defined types of the template data as mutable

fn get_mutable_template_data_global_variables(&mut self) -> &mut Vec<Variable>

Returns the global variables of the template data as mutable

fn get_mutable_template_data_functions(&mut self) -> &mut Vec<Function>

Returns the functions of the template data as mutable

fn get_template(&self, template_type: TemplateType) -> Template

Returns the template for the defined language based on the given template type